further refactoring of the code
