+++ analysis/analyze_fcts.py
@@ -25,106 +25,7 @@
 #
 ################################################################################
 
-import os
 import numpy as np
-
-from ._omega import omega
-
-def average_corr_fct(data, nbcfg, T):
-    """Average over the set of correlation functions.
-
-    Args:
-        data: The data to average over in a numpy array.
-        nbcfg: The number of configurations.
-        T: The time extent of the data.
-
-    Returns:
-        A numpy array averaged over one axis.
-    """
-    average = np.zeros((T))
-    for _t in range(T):
-        average[int(_t)] = np.average(data[_t*nbcfg:(_t+1)*nbcfg])
-    return average
-
-def simple_ratio(d1, d2, d3):
-    """Calculates a simple ratio of three data sets.
-
-    Calculates d1(t)/(d2(t)*d3(t)).
-    Assumes that all data sets are numpy arrays with two axis, the first being
-    the bootstrap number and the second being the time.
-    
-    Args:
-        d1, d2, d3: The three data sets.
-
-    Returns:
-        The ratio, its mean and its standard deviation.
-    """
-    # create array from dimensions of the data
-    ratio = np.zeros((d1.shape[0], d1.shape[1]))
-    for _s in range(d1.shape[0]):
-        for _t in range(d1.shape[1]):
-            # calculate ratio
-            ratio[_s,_t] = d1[_s,_t]/(d2[_s,_t]*d3[_s,_t])
-    # TODO(CJ): test if the following give the same result
-    # ratio = d1/(d2*d3)
-    # get mean and standard deviation
-    mean, err = return_mean_corr(ratio)
-    return ratio, mean, err
-
-def simple_ratio_subtract(d1, d2, d3):
-    """Calculates a simple ratio of three data sets, combining two time slices.
-
-    Calculates [d1(t)-d1(t+1)]/[(d2(t)*d3(t))-(d2(t+1)*d3(t+1))].
-    Assumes that all data sets are numpy arrays with two axis, the first being
-    the bootstrap number and the second being the time. The time extend is
-    reduced by one, as the ratio cannot be calculated on the last slice.
-    
-    Args:
-        d1, d2, d3: The three data sets.
-
-    Returns:
-        The ratio, its mean and its standard deviation.
-    """
-    # create array from dimensions of the data
-    ratio = np.zeros((d1.shape[0], d1.shape[1]-1))
-    for _s in range(d1.shape[0]):
-        for _t in range(d1.shape[1]-1):
-            # calculate ratio
-            ratio[_s,_t] = (d1[_s,_t] - d1[_s,_t+1]) / ((d2[_s,_t]*d3[_s,_t]) -
-                            (d2[_s,_t+1]*d3[_s,_t+1]))
-    # get mean and standard deviation
-    mean, err = return_mean_corr(ratio)
-    return ratio, mean, err
-
-def ratio(d1, d2, d3, dE):
-    """Calculates a ratio of three data sets, combining two time slices and
-       the energy difference between the energy levels.
-
-    Calculates [d1(t)-d1(t+1)]/[(d2(t)*d3(t))-(d2(t+1)*d3(t+1))].
-    Assumes that all data sets are numpy arrays with two axis, the first being
-    the bootstrap number and the second being the time. The time extend is
-    reduced by one, as the ratio cannot be calculated on the last slice.
-    
-    Args:
-        d1, d2, d3: The three data sets.
-        dE: Energy difference between the the data sets d2 and d3.
-
-    Returns:
-        The ratio, its mean and its standard deviation.
-    """
-    # create array from dimensions of the data
-    ratio = np.zeros((d1.shape[0], d1.shape[1]-1))
-    for _s in range(d1.shape[0]):
-        for _t in range(d1.shape[1]-1):
-            # calculate numerator and denominator first
-            num=(d1[_s,_t] - d1[_s,_t+1] * np.exp(dE * (_t+1))) * np.exp(-dE*_t)
-            den=(d2[_s,_t] * d3[_s,_t] - d2[_s,_t+1] * d3[_s,_t+1] *
-                 np.exp(dE*(_t+1))) * np.exp(-dE*_t)
-            # calculate ratio
-            ratio[_s,_t] = num/den
-    # get mean and standard deviation
-    mean, err = return_mean_corr(ratio)
-    return ratio, mean, err
 
 def compute_derivative(data):
     """Computes the derivative of a correlation function.
@@ -184,112 +85,7 @@ def compute_mass(data, usecosh=True):
     mean, err = return_mean_corr(mass)
     return mass, mean, err
 
-def calculate_q(E, mpi, L, lattice=False):
-    """Calculates q.
-
-    Calculates the difference in momentum between interaction and non-
-    interacting systems. The energy must be the center of mass energy.
-
-    Args:
-        E: The energy values for the bootstrap samples.
-        mpi: The pion mass of the lattice.
-        L: The spatial extent of the lattice.
-        lattice: Use the lattice relation, see arxiv:1011.5288.
-
-    Returns:
-        An array of the q^2 values
-    """
-    # create array for results
-    q2 = np.zeros((E.shape))
-    # lattice dispersion relation, see arxiv:1011.5288
-    if lattice:
-        q2 = (np.arcsin(np.sqrt((np.cosh(E*0.5)-np.cosh(mpi))*0.5))*float(L)/
-              np.pi)**2
-    # continuum dispersion relation
-    else:
-        q2 = (0.25*E*E - mpi*mpi) * (float(L) / (2. * np.pi))**2
-    return q2
-
-def calculate_delta(q2, gamma=None, d=np.array([0., 0., 0.]), prec=10e-6,
-                    verbose=0):
-    """Calculates the phase shift using Luescher's Zeta function.
-
-    Most arguments are for the Zeta function. For each q2 a gamma is needed.
-    WARNING: The momentum vectors d are compared to hardcoded momentum vectors
-    because of the zeta function calculation, which was derived for exactly
-    these momentum vectors. Make sure to use the right ones.
-    The names of the Irreps are the same as in M. Goeckeler et al., Phys. Rev.
-    D 86, 094513 (2012).
-
-    Args:
-        q2: The momentum shift squared.
-        gamma: The Lorentz factor for moving frames.
-        d: The total three momentum of the system.
-        prec: The precision of the Zeta function calculation.
-        verbose: The amount of information printed to screen.
-
-    Returns:
-        An array of the phase shift and tan(delta).
-    """
-    # create array for results
-    delta=np.zeros(q2.shape)
-    tandelta=np.zeros(q2.shape)
-    _gamma = gamma
-    if _gamma == None:
-        _gamma = np.ones(q2.shape)
-    # init calculation
-    _pi3 = np.pi**3
-    _num = _gamma * np.sqrt(_pi3 * q2)
-    #CMF
-    if np.array_equal(d, np.array([0., 0., 0.])):
-        # Irrep. T_1
-        #_z1 = zeta.Zp(q2, _gamma, 0, 0, d, 1., prec, verbose).real
-        #tandelta = np.sqrt( _pi3 * q2) / _z1.real
-        #delta = np.arctan2(np.sqrt( _pi3 * q2), _z1.real)
-        _den = omega(q2, _gamma, 0, 0, d, exFac=True).real
-        tandelta = _num / _den
-        delta = np.arctan2( _num, _den)
-    # MF1
-    elif np.array_equal(d, np.array([0., 0., 1.])):
-        # Irrep. A_1
-        #_z1 = zeta.Zp(q2, _gamma, 0, 0, d, 1., prec, verbose).real
-        #_z2 = zeta.Zp(q2, _gamma, 2, 0, d, 1., prec, verbose).real
-        #_num = _gamma * np.sqrt(_pi3 * q2)
-        #_den = (_z1 + (2. / (np.sqrt(5) * q2)) * _z2).real
-        _den = omega(q2, _gamma, 0, 0, d, exFac=True).real + \
-               2 * omega(q2, _gamma, 2, 0, d, exFac=True).real
-        tandelta = _num / _den
-        delta = np.arctan2( _num, _den)
-    # MF2
-    elif np.array_equal(d, np.array([1., 1., 0.])):
-        # Irrep. A_1
-        #_z1 = zeta.Zp(q2, _gamma, 0, 0, d, 1., prec, verbose).real
-        #_z2 = zeta.Zp(q2, _gamma, 2, 0, d, 1., prec, verbose).real
-        #_z3 = zeta.Zp(q2, _gamma, 2, 2, d, 1., prec, verbose).imag
-        #_num = _gamma * np.sqrt(_pi3 * q2)
-        #_den = (_z1 - (1. / (np.sqrt(5) * q2)) * _z2 + ( np.sqrt(6./5.) /
-        #        q2) * _z3 ).real
-        _den = omega(q2, _gamma, 0, 0, d, exFac=True).real -\
-               omega(q2, _gamma, 2, 0, d, exFac=True).real +\
-               np.sqrt(6) * omega(q2, _gamma, 2, 2, d, exFac=True).imag
-        tandelta = _num / _den
-        delta = np.arctan2( _num, _den)
-    # MF3
-    elif np.array_equal(d, np.array([1., 1., 1.])):
-        # Irrep. A_1
-        #_z1 = zeta.Zp(q2, _gamma, 0, 0, d, 1., prec, verbose).real
-        #_z2 = zeta.Zp(q2, _gamma, 2, 2, d, 1., prec, verbose).imag
-        #_num = _gamma * np.sqrt(_pi3 * q2)
-        #_den = (_z1 - ( 2. * np.sqrt(6./5.) / q2) * _z2 ).real
-        _den = omega(q2, _gamma, 0, 0, d, exFac=True).real -\
-               2 * np.sqrt(6) * omega(q2, _gamma, 2, 2, d, exFac=True).imag
-        tandelta = _num / _den
-        delta = np.arctan2( _num, _den)
-    else:
-        print("for the current vector d delta is not implemented")
-    return delta, tandelta
-
-def return_mean_corr(data, axis=0):
+def calc_error(data, axis=0):
     """Calculates the mean and standard deviation of the correlation function.
 
     Args:
