+++ xu_paper
@@ -102,7 +102,7 @@ def main():
     # calculate q2
     for _l in range(data.shape[0]):
         for _i in range(data.shape[2]):
-            q2[_l,:,_i] = ana.calculate_q(Ecm[_l,:,_i], mpi[_l], L[_l], True)
+            q2[_l,:,_i] = ana.calc_q2(Ecm[_l,:,_i], mpi[_l], L[_l], True)
             p[_l,:,_i] = np.sqrt(q2[_l,:,_i])*2*np.pi/float(L[_l])
     #print(q2)
     #print(p)
